Extracted download pprof into a utility func
